+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 In order to read more about upgrading and BC breaks have a look at the [UPGRADE Document](UPGRADE.md).
 
+## 1.0.22
+
++ Added new admin database table trait methods.
++ Ensure fixture data can be empty on various methods.
+
 ## 1.0.21 (17. September 2019)
 
 + Added new PageScope test system for CMS Pages.